Sign In
- Generated token
- Added tests
- Separated endpoint's tests in different files
