+++ app/errors.js
@@ -15,8 +15,5 @@ exports.invalidParamsError = message => internalError(message, exports.INVALID_P
 exports.UNIQUE_EMAIL_ERROR = 'unique_email_error';
 exports.uniqueEmailError = message => internalError(message, exports.UNIQUE_EMAIL_ERROR);
 
-exports.UNREGISTERED_EMAIL_ERROR = 'unregistered_email_error';
-exports.unregisteredEmailError = message => internalError(message, exports.UNREGISTERED_EMAIL_ERROR);
-
 exports.CREDENTIALS_ERROR = 'credentials_error';
 exports.credentialsError = message => internalError(message, exports.CREDENTIALS_ERROR);
